# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Travel Agency Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /flights/ksa-groups:
    get:
      operationId: listKsaGroups
      tags: [flights]
      summary: List KSA one way groups
      parameters:
        - name: airline
          in: query
          schema:
            type: ["string", "null"]
        - name: departureDate
          in: query
          schema:
            type: ["string", "null"]
        - name: minSeats
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: KSA groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlightGroup"
  /flights/ksa-groups/{id}:
    get:
      operationId: getKsaGroup
      tags: [flights]
      summary: Get KSA group detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: KSA group detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightGroup"
  /flights/uae-groups:
    get:
      operationId: listUaeGroups
      tags: [flights]
      summary: List UAE one way groups
      parameters:
        - name: airline
          in: query
          schema:
            type: ["string", "null"]
        - name: departureDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: UAE groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlightGroup"
  /flights/uae-groups/{id}:
    get:
      operationId: getUaeGroup
      tags: [flights]
      summary: Get UAE group detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: UAE group detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightGroup"
  /flights/all-groups:
    get:
      operationId: listAllGroups
      tags: [flights]
      summary: List all flight groups
      parameters:
        - name: airline
          in: query
          schema:
            type: ["string", "null"]
        - name: destination
          in: query
          schema:
            type: ["string", "null"]
        - name: departureDate
          in: query
          schema:
            type: ["string", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: All groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlightGroup"
  /flights/umrah-tickets:
    get:
      operationId: listUmrahTickets
      tags: [flights]
      summary: List umrah tickets
      parameters:
        - name: airline
          in: query
          schema:
            type: ["string", "null"]
        - name: departureDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Umrah tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlightGroup"
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/hold:
    post:
      operationId: holdBooking
      tags: [bookings]
      summary: Put booking on hold (2hr limit)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking on hold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/confirm:
    post:
      operationId: confirmBooking
      tags: [bookings]
      summary: Confirm a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/cancel:
    post:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /ledger:
    get:
      operationId: listLedger
      tags: [ledger]
      summary: List ledger entries
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LedgerEntry"

    post:
      operationId: createLedgerEntry
      tags: [ledger]
      summary: Create ledger entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LedgerEntryInput"
      responses:
        "201":
          description: Ledger entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerEntry"
...
components:
  schemas:
...
    Package:
      type: object
      required: [id, name, duration, type, price, hotel, airline, departureDate, returnDate, seatsAvailable, description, inclusions]
      properties:
        id:
          type: integer
        name:
          type: string
        duration:
          type: integer
        type:
          type: string
        price:
          type: number
        hotel:
          type: string
        hotelRating:
          type: ["integer", "null"]
        airline:
          type: string
        departureDate:
          type: string
        returnDate:
          type: string
        seatsAvailable:
          type: integer
        description:
          type: string
        inclusions:
          type: array
          items:
            type: string
        makkahNights:
          type: ["integer", "null"]
        madinahNights:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
    FlightGroup:
      type: object
      required: [id, flightNumber, airline, origin, destination, departureDate, departureTime, arrivalTime, seats, price, type, pnr, class]
      properties:
        id:
          type: integer
        flightNumber:
          type: string
        airline:
          type: string
        airlineCode:
          type: string
        origin:
          type: string
        originCode:
          type: string
        destination:
          type: string
        destinationCode:
          type: string
        departureDate:
          type: string
        departureTime:
          type: string
        arrivalTime:
          type: string
        duration:
          type: string
        seats:
          type: integer
        seatsAvailable:
          type: integer
        price:
          type: number
        type:
          type: string
        pnr:
          type: string
        class:
          type: string
        baggage:
          type: ["string", "null"]
        meal:
          type: ["boolean", "null"]
        refundable:
          type: ["boolean", "null"]
        stops:
          type: ["integer", "null"]
        aircraft:
          type: ["string", "null"]
    BookingInput:
      type: object
      required: [passengersInfo, flightGroupId, contactEmail, contactPhone, paymentMethod]
      properties:
        flightGroupId:
          type: ["integer", "null"]
        packageId:
          type: ["integer", "null"]
        passengersInfo:
          type: array
          items:
            $ref: "#/components/schemas/PassengerInfo"
        contactEmail:
          type: string
        contactPhone:
          type: string
        paymentMethod:
          type: string
        agentId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    BookingUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
    PassengerInfo:
      type: object
      required: [firstName, lastName, passportNumber, nationality, dateOfBirth]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        passportNumber:
          type: string
        nationality:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
    Booking:
      type: object
      required: [id, bookingRef, status, totalAmount, createdAt, passengersInfo, contactEmail, contactPhone]
      properties:
        id:
          type: integer
        bookingRef:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        holdExpiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
        flightGroup:
          $ref: "#/components/schemas/FlightGroup"
        package:
          $ref: "#/components/schemas/Package"
        passengersInfo:
          type: array
          items:
            $ref: "#/components/schemas/PassengerInfo"
        contactEmail:
          type: string
        contactPhone:
          type: string
        agentId:
          type: ["integer", "null"]
        agentName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        paymentMethod:
          type: string
...
    LedgerEntryInput:
      type: object
      required: [type, amount, description]
      properties:
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        bookingId:
          type: ["integer", "null"]
        bankId:
          type: ["integer", "null"]
    LedgerEntry:
      type: object
      required: [id, type, amount, description, balance, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        balance:
          type: number
        bookingId:
          type: ["integer", "null"]
        bankId:
          type: ["integer", "null"]
        createdAt:
          type: string
...