# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Agency Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { PassengerInfo } from "./passengerInfo";

export interface BookingInput {
  /** @nullable */
  flightGroupId: number | null;
  /** @nullable */
  packageId?: number | null;
  passengersInfo: PassengerInfo[];
  contactEmail: string;
  contactPhone: string;
  paymentMethod: string;
  /** @nullable */
  agentId?: number | null;
  /** @nullable */
  notes?: string | null;
}