# lib/api-zod/src/generated/types/bookingUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Agency Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingUpdate {
  status?: string;
  notes?: string;
}